#2889 - Refactored e2e tests to be a bit more readable and isolated
