+++ test/protractorConf.js
@@ -27,6 +27,8 @@ var config = {
   specs: [
           'tests/landing-page.js',
           'tests/notebook.js',
+          'tests/code-cell.js',
+          'tests/cell-menu.js',
           'tests/language-manager.js',
           'tests/bugs.js'
   ]
